Changed the way simulations are run, now all examples are models. Have moved all into the models module.
Removed lots of duplicate code since all examples are now inheriting functions of models - so components need to be set up but no longer have to do all the thread starting/joining.

How about Posidonius as a name...?

Started the python parsing code in codefile.py Was really hoping to avoid having to import the module with the python interpretter that is running the GUI - this would allow different versions of Python to be used for the GUI/simulations. But importing each module and using Pythons strong code introspection seems the best way of finding how many inputs or outputs a block has.

Added a stub class DisplayActor to actor.py which serves as a abstract base class for all plotters, required for the TK threading when running simulations with plots - see model.py run function for more info.

Made a composite actors submodule of models. This currently has the two sin source implementations. The two seperate older implementaitons of ctsin and dtsin are still present in models/actors but are not imported automatically into the actors submodule.